metafinder/requirements.txt
pikepdf==6.2.0
Pillow==9.2.0
ffmpeg-python==0.2.0
tinytag==1.8.1
colorama==0.4.4
